pyproject.toml
[project]
name = "domainhacks"
version = "0.1.0"
description = "Add a short description here"
authors = [
    { name = "Kuranai", email = "mail@kuranai.de" }
]
dependencies = [
    "whoisdomain>=1.20230525.9",
    "python-dotenv>=1.0.0",
    "psycopg[binary]>=3.1.9",
    "dnspython>=2.3.0",
    "requests>=2.31.0",
    "pandas>=2.0.3",
    "currencyconverter>=0.17.9",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=23.3.0",
    "ruff>=0.0.275",
]

[tool.hatch.metadata]
allow-direct-references = true
